Move payload parsing to common class
